requirements.txt
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
comm==0.2.2
confection==0.1.5
cymem==2.0.8
debugpy==1.8.6
decorator==5.1.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
filelock==3.16.1
fsspec==2024.9.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
langcodes==3.4.1
language_data==1.2.0
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
rich==13.9.1
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy_annotator==2.1.4
SpeechRecognition==3.10.4
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.3
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
widgetsnbextension==4.0.13
wrapt==1.16.0
